Add license, split docs, sandbox plugins
