+++ .env.example
@@ -9,7 +9,7 @@ BOT_TOKEN_MIXTRAL=
 BOT_TOKEN_DEEPSEEK=
 
 # Telegram webhook secret
-TG_SECRET=
+TG_SECRET=your-telegram-secret
 
 # Relay configuration
 RELAY_URL=http://localhost:5000/relay
