+++ docs/MODULES.md
@@ -1,0 +1,9 @@
+# Core Modules
+
+- `plugin_framework.py` – manages plug-ins for gestures and actions.
+- `workflow_controller.py` – runs workflows and reflex experiments.
+- `presence_ledger.py` – immutable memory of blessings and federation events.
+- `emotion_dashboard.py` – real-time emotion tracking interface.
+- `reflection_stream.py` – logging for autonomous reflections and diagnostics.
+
+See other files in this repository and `docs/README_FULL.md` for a detailed tour.
