+++ logs/federation_log.jsonl
@@ -1,3 +1,5 @@
 {"timestamp": "2025-06-01T01:00:16.954728", "peer": "remote", "email": "", "message": "imported logs", "ritual": "Federation blessing invoked and logged."}
 {"timestamp": "2025-06-01T14:02:00.200435", "peer": "remote", "email": "", "message": "import payload", "ritual": "Federation blessing recorded."}
 {"timestamp": "2025-06-01T14:02:00.200538", "peer": "remote", "email": "", "message": "imported logs", "ritual": "Federation blessing recorded."}
+{"timestamp": "2025-06-01T20:24:09.608850", "peer": "remote", "email": "", "message": "import payload", "ritual": "Federation blessing recorded."}
+{"timestamp": "2025-06-01T20:24:09.609001", "peer": "remote", "email": "", "message": "imported logs", "ritual": "Federation blessing recorded."}
