+++ logs/support_log.jsonl
@@ -22,3 +22,8 @@
 {"timestamp": "2025-06-01T14:02:00.979228", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
 {"timestamp": "2025-06-01T14:02:00.979287", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
 {"timestamp": "2025-06-01T14:02:00.979359", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
+{"timestamp": "2025-06-01T20:24:10.039765", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
+{"timestamp": "2025-06-01T20:24:10.353621", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
+{"timestamp": "2025-06-01T20:24:10.353712", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
+{"timestamp": "2025-06-01T20:24:10.353789", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
+{"timestamp": "2025-06-01T20:24:10.353847", "supporter": "anon", "message": "analysis blessing: memory record", "amount": "", "ritual": "Sanctuary blessing acknowledged and remembered."}
