+++ plugin_dashboard.py
@@ -1,4 +1,4 @@
-from flask import Flask, jsonify, request
+from flask_stub import Flask, jsonify, request
 import plugin_framework as pf
 import trust_engine as te
 from admin_utils import require_admin_banner
